days 9 and 10: use Eloquent ORM to get/set data and use factories to scaffold data
Eloquent ORM. Object relational mapper. Maps a table row to an object in php
code.

Normally, you will make a model with `php artisan make:model` and sometimes
also create a migration alongside that model with `php artisan make:model -r`

php artisan tinker is a repl which can be used to write ad-hoc php code.

We can create factories that produce fake data. We can include functions inside
of these factories that help to modify the state for particular configurations
of data. By default the UsersFactory.php has an `unverified()` function for
creating a user that doesn't have a verified email address.

A factory can reference a different model to create a different entry and use
that as a foreign key

Factories are good at scaffolding data for a local environment