Day 12: how to do pivot (linking) tables
we can use the `belongsToMany` function to establish a many-to-many
relationship using a pivot table.

we can add foreign constraints to prevent orphans in the pivot table.

By default sqlite doesn't enforce constraints, but Laravel does. So this means
that if a record is deleted outside of Laravel, then the cascading delete won't
happen unless that option is turned on. But if that delete happens in Laravel,
then the cascade will happen.