day 14: use pagination on the jobs page
instead of using `get()`, using `paginate()` enables pagination. Then then you
can do `{{ $items->links() }}` to reder the paginator that is already styled
with tailwind out of the box.

The paginator code isn't the repo by default, but it can be added by running
`php artisan vendor:publish` and selecting the paginator. This copies that code
from the dependencies from the vendor folder (where packages live) into the
current project where it can be configured.

`simplePaginate()` or `cursorPaginate()` can be used. Cursor pagination is most
performant, but the url isn't as friendly. Cursor pagination is good for things
where the url friendliness is not as important like infinite scrolling.