day 15: use database seeders to automatically create fake data
So far we have only populated the database semi-manually using
`php artisan tinker` and running something like
`App\Models\Job::factory(100)->create()`. A way to do this in a more automated
way is with database seeders.

There is a default seeder that can be run with `php artisan db:seed` or
`php artisan migrate:fresh --seed` if the database should be newly created.

Other seeders can be created using `php artisan make:seeder` for alternate
scenarios, such as tests that need a particular data configuration.

Rule of thumb is to keep everything in the same seeder file, keeping it simple,
until there is a need to create additional seeders.

The non-default seeder can be run with
`php artisan migrate:fresh --seed=OtherSeeder` where "OtherSeeder" is the name
of that seeder.

Seeders typically use factories, but the don't need to used them. It just so
happens that factories are a very convenient way to create a lot of fake data.
Reaching for eloquent directly to create a record with hard-coded values is
also ok.