day 16: do a form post
The sequence of declared routes matters. Wildcards should be declared after
other things.

There are naming conventions for breaking up the views directory into
additional directories; index is for the base route, show is for a specific one
of that base route (think "/jobs/123").

The convention for a creation form is to have a form at "/jobs/create" for
example, and from there do a form POST to action="/jobs".

Laravel throws a 419 error response when a form is submitted (at least with
POST method) due to danger of cross site request forgery (CSRF). In Laravel
blade we can use the @csrf directive to take care of adding a hidden element
with the CSRF token.

fillable fields protection can be disabled by setting an empty guarded field in the model.