day 19: clean up the routes file
Instead of doing a find on the id in every request, using the model type is a
good way to go.

running `php artisan route:list --except-vendor` will list all the routes of
our app

`Route::view` and `Route::resource` are some shortcuts to simply do a get ->
view and register an entire controller respectively. Remember that there are
conventional names that are used for the functions in a controller.