days 20 & 21: go over starter kits, and build register and login forms
Laravel has starter kit scaffolding that can be used with all sorts of stuff
built in including authentication. Use them with `laravel new <app-name>`

The breeze starter kit does authentication checks for the user dashboard using
middleware.

These starter kits are a great resource to study.

Question: Instead of componentizing clusters of html, wouldn't it make sense to
use tailwind's `@apply` to create semantic class names instead? That way, we're
not introducing seemingly new elements that abstract over html. One of the
selling ideas behind tailwind is that it is difficult to create meaningful
class names. Don't we still need to create meaningful names when we
componentize stuff? Components facilitate reuse. If we favor pushing it all
down to css, we can continue to just write html without annoying abstractions
and we still have a degree of code reuse using the semantic css class names.

It still might make sense to componentize where there is a large amount of code
reuse, like an entire widget, but for something that wraps a single element or
a handful of elements, just make a semantic class name an copy-and-paste that
junk. That's just my opinion. Maybe uninformed.