day 22: finish registration, login, and log out.
Using 'confirmed' in validation requires that an associated
`<name>_confirmation` is present in the request.

Laravel automatically has user password hashing set up. It uses a function
called `casts` in the User model which says specifies that the password should
be hashed.

Always use a post request to log out the user so other websites can't use a
link (get) to log you out.

Should review rate limiting for logging in.