day 23: implement authorization
`is` is a useful method on models which can be used to check for auth.

Gate is a good batteries-included way to do auth. The gate definitions can be
put in the AppServiceProvider so they can be referenced from any controller.
Also it can be referenced from blade templates with the `@can` or `@cannot`
directives.

In order to avoid pasting the authorization gate check into every function of a
controller, we can use middleware to handle that for us.

Instead of using a Gate, we can use a policy. For non-trivial applications, use
a policy.