day 24: sending emails
The default mailer is 'log' which outputs emails to a log file. The from
address can be configured from the env file, or for specific mail situations it
can be configured in those mail files which are created using `php artisan
make:mail`

Mail trap is mailer for which one can sign up for free to test sending emails.