day 25: add queues, jobs, and workers
Users shouldn't need to wait for emails to be delivered. Instead of sending
them synchronously, we can queue them up. The default queue uses the database,
which is good enough for a while.

When something is queued, it can be processed with `php artisan queue:work`.
When something is queued, we need one or more workers to process those things
on the queue.

work can be queued with a delay to it's execution.

A worker needs to be restarted if a job changes, or else the worker will take
the out of date actions.