day 26: css and js bundling
Css needs to be compiled by Vite. It can be referenced in a blade template by
using the `@vite` directive. This has hot module reloading.

Also the javascript in the `resources` directory is bundled the same way and
can be included with the `@vite` directive.

Setup tailwind the right way.

If building an spa, import the css classes from the bundle directly.